+++ src/aks-preview/azext_aks_preview/_params.py
@@ -686,6 +686,12 @@ def load_arguments(self, _):
             help="enable vertical pod autoscaler for cluster",
         )
         c.argument(
+            "enable_addon_autoscaling",
+            action="store_true",
+            is_preview=True,
+            help="enable addon autoscaling for cluster",
+        )
+        c.argument(
             "enable_node_restriction",
             action="store_true",
             is_preview=True,
@@ -1093,6 +1099,18 @@ def load_arguments(self, _):
             action="store_true",
             is_preview=True,
             help="disable vertical pod autoscaler for cluster",
+        )
+        c.argument(
+            "enable_addon_autoscaling",
+            action="store_true",
+            is_preview=True,
+            help="enable addon autoscaling for cluster",
+        )
+        c.argument(
+            "disable_addon_autoscaling",
+            action="store_true",
+            is_preview=True,
+            help="disable addon autoscaling for cluster",
         )
         c.argument(
             "cluster_snapshot_id",