+++ src/aks-preview/azext_aks_preview/managed_cluster_decorator.py
@@ -1974,6 +1974,58 @@ class AKSPreviewManagedClusterContext(AKSManagedClusterContext):
         """
         return self._get_disable_vpa(enable_validation=True)
 
+    def _get_enable_addon_autoscaling(self, enable_validation: bool = False) -> bool:
+        """Internal function to obtain the value of enable_addon_autoscaling.
+        This function supports the option of enable_addon_autoscaling. When enabled, if both enable_addon_autoscaling and disable_addon_autoscaling are
+        specified, raise a MutuallyExclusiveArgumentError.
+        :return: bool
+        """
+        # Read the original value passed by the command.
+        enable_addon_autoscaling = self.raw_param.get("enable_addon_autoscaling")
+
+        # This parameter does not need dynamic completion.
+        if enable_validation:
+            if enable_addon_autoscaling and self._get_disable_addon_autoscaling(enable_validation=False):
+                raise MutuallyExclusiveArgumentError(
+                    "Cannot specify --enable-addon-autoscaling and --disable-addon-autoscaling at the same time."
+                )
+
+        return enable_addon_autoscaling
+
+    def get_enable_addon_autoscaling(self) -> bool:
+        """Obtain the value of enable_addon_autoscaling.
+        This function will verify the parameter by default. If both enable_addon_autoscaling and disable_addon_autoscaling are specified, raise
+        a MutuallyExclusiveArgumentError.
+        :return: bool
+        """
+        return self._get_enable_addon_autoscaling(enable_validation=True)
+
+    def _get_disable_addon_autoscaling(self, enable_validation: bool = False) -> bool:
+        """Internal function to obtain the value of disable_addon_autoscaling.
+        This function supports the option of enable_addon_autoscaling. When enabled, if both enable_addon_autoscaling and disable_addon_autoscaling are specified,
+        raise a MutuallyExclusiveArgumentError.
+        :return: bool
+        """
+        # Read the original value passed by the command.
+        disable_addon_autoscaling = self.raw_param.get("disable_addon_autoscaling")
+
+        # This option is not supported in create mode, hence we do not read the property value from the `mc` object.
+        # This parameter does not need dynamic completion.
+        if enable_validation:
+            if disable_addon_autoscaling and self._get_enable_addon_autoscaling(enable_validation=False):
+                raise MutuallyExclusiveArgumentError(
+                    "Cannot specify --enable-addon-autoscaling and --disable-addon-autoscaling at the same time."
+                )
+
+        return disable_addon_autoscaling
+
+    def get_disable_addon_autoscaling(self) -> bool:
+        """Obtain the value of disable_addon_autoscaling.
+        This function will verify the parameter by default. If both enable_addon_autoscaling and disable_addon_autoscaling are specified, raise a MutuallyExclusiveArgumentError.
+        :return: bool
+        """
+        return self._get_disable_addon_autoscaling(enable_validation=True)
+
     def get_ssh_key_value_for_update(self) -> Tuple[str, bool]:
         """Obtain the value of ssh_key_value for "az aks update".
 
@@ -2770,6 +2822,9 @@ class AKSPreviewManagedClusterCreateDecorator(AKSManagedClusterCreateDecorator):
                 self.models.ManagedClusterWorkloadAutoScalerProfileKeda(enabled=True)  # pylint: disable=no-member
             )
 
+        # set up addon autoscaling
+        mc = self.set_up_addon_autoscaling(mc)
+
         return mc
 
     def set_up_custom_ca_trust_certificates(self, mc: ManagedCluster) -> ManagedCluster:
@@ -2827,6 +2882,25 @@ class AKSPreviewManagedClusterCreateDecorator(AKSManagedClusterCreateDecorator):
                 )
             else:
                 mc.workload_auto_scaler_profile.vertical_pod_autoscaler.enabled = True
+        return mc
+    
+    def set_up_addon_autoscaling(self, mc: ManagedCluster) -> ManagedCluster:
+        """Set up workload auto-scaler vertical pod autsocaler profile for the ManagedCluster object.
+
+        :return: the ManagedCluster object
+        """
+        self._ensure_mc(mc)
+
+        if self.context.get_enable_addon_autoscaling():
+            if mc.workload_auto_scaler_profile is None:
+                mc.workload_auto_scaler_profile = self.models.ManagedClusterWorkloadAutoScalerProfile()
+            if mc.workload_auto_scaler_profile.vertical_pod_autoscaler is None:
+                mc.workload_auto_scaler_profile.vertical_pod_autoscaler = self.models.ManagedClusterWorkloadAutoScalerProfileVerticalPodAutoscaler()
+
+            # set enabled
+            mc.workload_auto_scaler_profile.vertical_pod_autoscaler.enabled = True
+            mc.workload_auto_scaler_profile.vertical_pod_autoscaler.addon_autoscaling = "Enabled"
+
         return mc
 
     def set_up_kube_proxy_config(self, mc: ManagedCluster) -> ManagedCluster:
@@ -3810,6 +3884,9 @@ class AKSPreviewManagedClusterUpdateDecorator(AKSManagedClusterUpdateDecorator):
                 self.models.ManagedClusterWorkloadAutoScalerProfileKeda(enabled=False)  # pylint: disable=no-member
             )
 
+        # update addon autoscaling
+        mc = self.update_addon_autoscaling(mc)
+
         return mc
 
     def update_custom_ca_trust_certificates(self, mc: ManagedCluster) -> ManagedCluster:
@@ -3948,6 +4025,34 @@ class AKSPreviewManagedClusterUpdateDecorator(AKSManagedClusterUpdateDecorator):
 
             # set disabled
             mc.workload_auto_scaler_profile.vertical_pod_autoscaler.enabled = False
+
+        return mc
+
+    def update_addon_autoscaling(self, mc: ManagedCluster) -> ManagedCluster:
+        """Update workload auto-scaler vertical pod auto-scaler profile for the ManagedCluster object.
+
+        :return: the ManagedCluster object
+        """
+        self._ensure_mc(mc)
+
+        if self.context.get_enable_addon_autoscaling():
+            if mc.workload_auto_scaler_profile is None:
+                mc.workload_auto_scaler_profile = self.models.ManagedClusterWorkloadAutoScalerProfile()
+            if mc.workload_auto_scaler_profile.vertical_pod_autoscaler is None:
+                mc.workload_auto_scaler_profile.vertical_pod_autoscaler = self.models.ManagedClusterWorkloadAutoScalerProfileVerticalPodAutoscaler()
+
+            # set enabled
+            mc.workload_auto_scaler_profile.vertical_pod_autoscaler.enabled = True
+            mc.workload_auto_scaler_profile.vertical_pod_autoscaler.addon_autoscaling = "Enabled"
+
+        if self.context.get_disable_addon_autoscaling():
+            if mc.workload_auto_scaler_profile is None:
+                mc.workload_auto_scaler_profile = self.models.ManagedClusterWorkloadAutoScalerProfile()
+            if mc.workload_auto_scaler_profile.vertical_pod_autoscaler is None:
+                mc.workload_auto_scaler_profile.vertical_pod_autoscaler = self.models.ManagedClusterWorkloadAutoScalerProfileVerticalPodAutoscaler()
+
+            # set disabled
+            mc.workload_auto_scaler_profile.vertical_pod_autoscaler.addon_autoscaling = "Disabled"
 
         return mc